# hopelink backend/backend/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "HopeLink API",
    "description": "Comprehensive API documentation for HopeLink Platform - A charity and volunteer management system",
    "version": "1.0.0",
    "contact": {
      "name": "HopeLink Support",
      "email": "support@hopelink.com"
    },
    "license": {
      "name": "MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3001/api/v1"
    },
    {
      "url": "https://localhost:3001/api/v1"
    }
  ],
...
  "paths": {
...
    "/me": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
...
    "/user/{userId}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "bearerAuth": {
        "type": "apiKey",
        "name": "Authorization",
        "in": "header",
        "description": "JWT Authorization header using the Bearer scheme. Example: \"Bearer {token}\""
      }
    }
  }
}